Userstory Crisis-Phase framework
The "res"-folder has been moved into the godot directionary for easy access in the godot-editor.
The Scene CrisisPhase was initiated with a simple collision border around the edge and a provisional Tilemap.
The avatar has been intitiated as a moveable (with keyboard controls w, a, s, d) character with a provisional sprite and a collision shape.