Update syntax-basics.md
Added actions, conditions and file structure.

There are some inconsistencies right now. These are related to the following sections:
- Condition: no [action] tag is used, the "and" is not used so far, instead conditions are just listed (-> what to use?)
- Progress day and Gain item: used like this in game right now, if I should use them differently, we can adapt here
- Variables, Karma, Friendship, Mood, Crisis Effect: no [action] tag is used, but mostly used in game like this right now, if I should use them differently, we can adapt here
- Crisis, Stage: have not been used in game yet and need to be added, therefore I used the [action] tag here

Currently, day dialogues are all in day.vrv etc., so the file structure is also not the one presented here.

Imho, we should define the API that will be used on the 21.11., so the dialogues work. Afterwards, we can adapt, if necessary. Right now, priority is making the dialogues work (or keeping them working). This is why I did not change the file structure yet. I'm open to change things here or update them later.